+++ src/test/java/racingcar/model/domain/CarsManagerTest.java
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 import racingcar.Application;
-import racingcar.car.model.CarDTO;
+import racingcar.car.model.dto.CarDTO;
 import racingcar.car.model.domain.CarsManager;
 
 import java.util.ArrayList;
